Add files via upload
Particularly, I implement A-B testing and utilize metrics such as the t-test, 3 way ANOVA( Analysis of Variance) to determine if the test is statistically significant and whether the null hypothesis is rejected. If so then our findings are valid statistically. The hypothesis for this experiment is: Are Married people more likely to be satisfied with their jobs compared to Single people? The purpose of this is to use data and statistical knowledge to explore these trends which may not always be conspicuous in order to help managers and employers better address these issues (if any)