Add files via upload
The purpose of the project was to showcase my capabilities accessing data through web scraping. Web scraping is simply exctracting data from websites.For this project I chose data from a website 'https://www.space.com/news' to obtain the main content of the news. Web scraping is an alternative to when data from the website's API is not readily available.The data can be used for various forms of analysis. Prior to extracting the data, I used the website's elements tab to visualize the different parts of the webpage and get a better understanding of the different elements. I then used the requests library to fetch the code from the website. I then carefully studied the data, examined the transitions between class names, styles, divisions, lists etc. I also used beautiful soup to parse the HTML and extract the content and also nmake sure the output code is well formatted. I created various functions to extract the specific elements I needed by using the distinct names and types. After all the analysis, I was able to obtain the specific information I needed  namely the author, date, headline, synopsis. This data can then be further analyzed for specific purposed, which wouldn't be possible if it were on the website.