Add files via upload
This data science project is about predicting the likelihood of customers of a telecom company to churn. "Customer Churn Rate" is defined as the percentage of customers that stopped using a company's product or service during a certain time frame. Analyzing the churn rate and underlying factors is important because in many cases its more costly to acquire new customers than it does to retain them. Additionally, the higher the customer retention, the more the profitability and business growth. It's also a good indicator of whether the product, service or team are succeeding.