Add files via upload
I sought this dataset because one of the things I am passionate about is improving employee and customer relations using data science(in this particular case employee data). This research focuses on observing several relationships between attrition at the workplace and other factors such as age, gender, education level, salary and so on. Particularly, I implement A-B testing and utilize metrics such as the t-test, 3 way ANOVA( Analysis of Variance) to determine if the test is statistically significant and whether the null hypothesis is rejected. If so then our findings are valid statistically. The hypothesis for this experiment is: Are Married people more likely to be satisfied with their jobs compared to Single people? The purpose of this is to use data and statistical knowledge to explore these trends which may not always be conspicuous in order to help managers and employers better address these issues (if any). It is also a classical example of application of A-B testing.The target variable is the atttition which has 3 classes; single, married and divorced. For my analysis, I mostly focused on single vs married employees. The dataset contains 1470 observations and 25 features.